Complete test for deals(transactions)
