+++ app/views/layouts/application.html.erb
@@ -6,7 +6,6 @@
     <%= csrf_meta_tags %>
     <%= csp_meta_tag %>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200" />
-    <!-- <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0-beta3/css/all.min.css" rel="stylesheet"> -->
     <%= stylesheet_link_tag "application", "splash", "auth", "category", "transactions", "data-turbo-track": "reload" %>
     <%= javascript_importmap_tags %>
   </head>
